requirements.txt
numpy
tensorflow
matplotlib
scikit-learn
opencv-python
openpyxl
tensorflow_datasets
pillow
scikit-image
